project: Apply more C++ paradigms to the code
- Use auto in places where code clarity is improved or identical.
- Replace trivial constructors and destructors with default.
- Use true random for random generation.
- Use std::string_view where it is valid to do so.
- Apply const where it is valid to do so.
- Use references where it is valid to do so.
- Manually optimize memory usage with std::move and std::copy.
- Opt for memory efficient containers where the size is known ahead of time.

+++ source/nvidia/vfx/nvidia-vfx-superresolution.cpp
@@ -134,7 +134,7 @@ void streamfx::nvidia::vfx::superresolution::set_scale(float scale)
 	scale = std::clamp<float>(scale, 1., 4.);
 
 	// Match to nearest scale.
-	double factor = find_closest_scale_factor(scale);
+	float factor = static_cast<float>(find_closest_scale_factor(scale));
 
 	// If anything was changed, flag the effect as dirty.
 	if (!::streamfx::util::math::is_close<float>(_scale, factor, 0.01f))
